Use percona repo
OSA defaults to pulling the xtrabackup deb from a url and installing it
locally. Unfortunately the version pulled is too old for compatibility
with rpc holland.

By setting use_percona_repo: yes, the percona repo will be added and the
latest version of the package will be installed from there.

The OSA team are working on upgrading the version installed by default
but suggested that it probably won't be backported to Mitaka or Newton,
so we need to carry this var for a while.

OSA Bug: https://bugs.launchpad.net/openstack-ansible/+bug/1653947
Connects: rcbops/u-suk-dev#825